Finally Made website Dynamic
Yay>> Finally made it dynamic.
You can add Data to it.
