+++ index.html
@@ -80,8 +80,8 @@
             <a href="https://fmoviesto.tv/">Fmovies</a>
             <a href="https://www9.123moviesto.to/">123movies.to</a>
             <a href="https://cataz.net/">Cataz</a>
-            <a href="https://www2.solarmovie.cr/movie/https://pdfstop.com/category/harry-potter-1/
-https://fmovies.to/home">Solar Movie</a>
+            <a href="https://www2.solarmovie.cr/movie/https://pdfstop.com/category/harry-potter-1/">
+            <a href="https://fmovies.to/home">Solar Movie</a>
             <a href="https://www.putlockers.do/movie/demolition-8084.html">Put Lockers</a>
             <a href="https://www.limetorrents.lol/browse-torrents/Movies/">Lime Torrents</a>
             <a href="https://rarbgaccessed.org/torrents.php?r=99481836">Rarbg</a>
