--- src/app2.js
@@ -1,2 +1,0 @@
-var list = document.getElementById("list");
-import websites from data.js;
