+++ src/index2.html
@@ -7,13 +7,53 @@
     <link rel="stylesheet" href="styles2.css">
 </head>
 <body>
-    <main>    
-        <div id="catlist"><span id="catName"></span>
-            <div id="list">
-                
+        <main>
+        <form class="search" action="https://search.brave.com/search" method="get" target="_blank">
+            <input class="sbar" type="text" name="q" placeholder="Search with Brave" onkeydown="return handleKeyDown(event);">
+        </form>
+        <h3 class="title-head">Applications</h3>
+        <div class="appstore">
+            <div class="applist">
+                <a target="_blank" href="https://github.com/">Github</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://chat.openai.com/">Chat GPT</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://www.youtube.com">Youtube</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://www.codewars.com/">CodeWars</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://frontendmasters.com/">Frontend Masters</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://codepen.io/your-work">CodePen</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://developer.mozilla.org/en-US/">MDN</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://www.freecodecamp.org/">FreeCodeCamp</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://www.hackerrank.com/">HackerRank</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="www.codewars.com">CodeWars</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://www.codecademy.com/">CodeAcademy</a>
+            </div>
+            <div class="applist">
+                <a target="_blank" href="https://devdocs.io/">Devdocs</a>
             </div>
         </div>
-    </main>
-</body>
+        <h3 class="title-head">Bookmarks</h3>
+        <div class="bmlist">
+
+        </body>
 <script src="data.js"></script>
+<script src="App.js"></script>
 </html>
