+++ src/styles2.css
@@ -1,5 +1,157 @@
-#list{
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
+:root{
+        --color-background: rgb(38, 50, 56);
+        --color-text-pri: #AABBC3;
+        --color-text-acc: #aeea00;
+    }
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
 }
+html, body {
+    width: 100%;
+    height: 100%;
+}
+body{
+    background-color: #fff;
+    background-color: var(--color-background);
+    color: var(--color-text-pri);
+    font-family: 'Exo 2', sans-serif;
+    font-family: 'Ubuntu', sans-serif;
+    font-size: 20px;
+}
+main{
+    width: 80%;
+    margin: auto;
+    padding-top: 50px;
+}
+
+.appstore a {
+    text-transform: uppercase;
+}
+a{
+    text-decoration: none;
+    color: inherit;
+    width: 100%;
+}
+a:hover{
+    color: var(--color-text-acc);
+    text-decoration: underline;
+}
+.search{
+    text-align: center;
+    /* padding: 50px; */
+}
+.sbar{
+    width: 80%;
+    font-size: 20px;
+    padding: 5px;
+    background-color: var(--color-background);
+    border-top: none;
+    border-bottom: 2px solid var(--color-text-acc);
+    border-right: none;
+    border-left: 1px;
+    color: var(--color-text-acc);
+}
+.title-head{
+    text-transform: uppercase;
+    padding: 25px 0px;
+    color: var(--color-text-acc);   
+    font-size: 25px;
+}
+.appstore{
+    width: 90%;
+    margin: auto;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    padding-bottom: 15px;
+    font-size: 20px;
+}
+.applist{
+    width: 22%;
+    padding: 8px;
+    margin: 15px 15px 0 15px;
+    border-radius: 7px;
+    box-shadow: 2px 2px 2px 2px rgba(29, 38, 43, 0.5);
+    
+}
+.applist:hover{
+    box-shadow: 0 2px 5px var(--color-text-acc);
+}
+.bmlist{
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    font-size: 16px;
+    text-align: center;
+}
+.bmcat{
+    padding: 10px;
+    text-transform: uppercase;
+    color: var(--color-text-acc);
+    display: flex;
+    flex-direction: row;
+}
+.bmcat::-webkit-scrollbar{
+    display: contents;
+}
+.sticky{
+    position: sticky;
+    top: 10px;
+    text-transform: uppercase;
+    color: var(--color-text-acc);
+    margin-bottom: 50px;
+}
+.siteList{
+    text-transform: uppercase;
+    text-align: start;
+    color: var(--color-text-pri);
+    line-height: 30px;
+}
+.siteList a {
+    padding: 15px 20px;
+}
+
+
+
+/* For adding mobile view */
+@media (max-width: 767px){
+.search{
+    padding: none;
+}
+.sbar {
+    width: 100%;
+}
+.applist {
+    width: 92%;
+}
+.bmlist {
+flex-direction: column;
+}
+.bmlist{
+padding: 10px;
+align-items: start;
+}
+.bmcat{
+padding: 5px;
+margin: 5px auto;
+width: 80%;
+box-shadow: 2px 2px 2px 2px rgba(29, 38, 43, 0.5);
+}
+}
+
+
+
+
+
+
+/*  
+ ignou & study
+
+    https://ignou.samarth.edu.in/index.php/site/login          
+    https://admissions.nic.in/IPUADM/Applicant/CommonProfile/Signup.aspx
+https://ipu.admissions.nic.in/
+
+                                   
+*/
